Use rust abi info to convert type
